Refactor create method to accept key=value params
The create method in console.py now accepts key=value parameters,
allowing for the creation of new objects with desired attributes. This
change enhances the flexibility and user-friendliness of creating new
instances, bolstering the functionality of the HBNB console. The
BaseModel class in base_model.py has also been updated with a set method
to handle attribute setting. This adjustment ensures that attribute
values are appropriately handled when being set, contributing to
improved data integrity. (Issue #123, #124)